DGEExact-class {edgeR}	R Documentation

differential expression of Digital Gene Expression data - class

Description

A simple list-based class for storing results of differential expression analysis for DGE data

Slots/List Components

Objects of this class contain the following list components:

table: data frame containing the log-concentration (i.e. expression level), the log-fold change in expression between the two groups/conditions and the exact p-value for differential expression, for each tag.

comparison: vector giving the two experimental groups/conditions being compared.

genes: a data frame containing information about each transcript (can be NULL).

Methods

This class inherits directly from class list so any operation appropriate for lists will work on objects of this class. DGEExact objects also have a show method.

Author(s)

Mark Robinson, Davis McCarthy

[Package edgeR version 2.4.3 Index]