DGEGLM-class {edgeR}	R Documentation

Digital Gene Expression Generalized Linear Model results - class

Description

A simple list-based class for storing results of a GLM fit to each tag/gene in a DGE dataset.

Slots/List Components

Objects of this class contain the following list components:

coefficients: matrix containing the coefficients computed from fitting the model defined by the design matrix to each gene/tag in the dataset.

df.residual: vector containing the residual degrees of freedom for the model fit to each tag/gene in the dataset.

deviance: vector giving the deviance from the model fit to each tag/gene.

design: design matrix for the full model from the likelihood ratio test.

offset: scalar, vector or matrix of offset values to be included in the GLMs for each tag/gene.

samples: data frame containing information about the samples comprising the dataset.

genes: data frame containing information about the genes or tags for which we have DGE data (can be NULL if there is no information available).

dispersion: scalar or vector providing the value of the dispersion parameter used in the negative binomial GLM for each tag/gene.

lib.size: vector providing the effective library size for each sample in the dataset.

weights: matrix of weights used in the GLM fitting for each tag/gene.

fitted.values: the fitted (expected) values–here they are counts–from the GLM for each tag/gene.

abundance: vector of gene/tag abundances (expression level), on the log2 scale, computed from the mean count for each gene/tag after scaling count by normalized library size.

Methods

This class inherits directly from class list so any operation appropriate for lists will work on objects of this class. DGEGLM objects also have a show method.

Author(s)

Davis McCarthy

[Package edgeR version 2.4.3 Index]