DGELRT-class {edgeR}	R Documentation

Digital Gene Expression Likelihood Ratio Test data and results - class

Description

A simple list-based class for storing results of a GLM-based differential expression analysis for DGE data, with evidence for differential expression assessed using a likelihood ratio test.

Slots/List Components

Objects of this class contain the following list components:

table: data frame containing the log-concentration (i.e. expression level), the log-fold change in expression between the two groups/conditions and the exact p-value for differential expression, for each tag.

coefficients.full: matrix containing the coefficients computed from fitting the full model (fit using glmFit and a given design matrix) to each gene/tag in the dataset.

coefficients.null: matrix containing the coefficients computed from fitting the null model to each gene/tag in the dataset. The null model is the model to which the full model is compared, and is fit using glmFit and dropping selected column(s) (i.e. coefficient(s)) from the design matrix for the full model.

design: design matrix for the full model from the likelihood ratio test.

...: if the argument y to glmLRT (which produces the DGELRT object) was itself a DGEList object, then the DGELRT will contain all of the elements of y, except for the table of counts and the table of pseudocounts.

Methods

This class inherits directly from class list so any operation appropriate for lists will work on objects of this class. DGELRT objects also have a show method.

Author(s)

Davis McCarthy

[Package edgeR version 2.4.3 Index]