DGEList-class {edgeR}	R Documentation

Digital Gene Expression data - class

Description

A simple list-based class for storing read counts from digital gene expression technologies and other important information for the analysis of DGE data.

Slots/List Components

Objects of this class contain (at least) the following list components:

counts: numeric matrix containing the read counts.

samples: data.frame containing the library size and group labels.

Methods

This class inherits directly from class list so any operation appropriate for lists will work on objects of this class. DGEList objects also have a show method.

Author(s)

Mark Robinson

See Also

DGEList

[Package edgeR version 2.4.3 Index]