approx.expected.info {edgeR}	R Documentation

Approximate Expected Information (Fisher Information)

Description

Using a linear fit (for simplicity), the expected information from the conditional log likelihood of the dispersion parameter of the negative binomial is calculated over all genes.

Usage

approx.expected.info(object, d, pseudo, robust = FALSE) 

Arguments

object
DGEList object containing the raw counts with (at least) elements counts (table of counts), group (vector indicating group) and lib.size (vector of library sizes)

d
numeric vector giving the delta parameter for negative binomial - phi/(phi+1) ; either of length 1 or of length equal to the number of tags/transcripts (i.e. number of rows of object$counts.

pseudo
numeric matrix of pseudocounts from output of estimateDispIter

robust
logical on whether to use a robust fit, default FALSE

Value

numeric vector of approximate values of the Fisher information for each tag/transcript (with length same as the number of rows of the original counts)

Author(s)

Mark Robinson

See Also

This function is used in the algorithm for estimating an appropriate amount of smoothing for the dipsersion estimates carried out by estimateSmoothing.

Examples

set.seed(0)
y<-matrix(rnbinom(40,size=1,mu=10),ncol=4)
d<-DGEList(counts=y,group=rep(1:2,each=2),lib.size=rep(c(1000:1001),2))
d<-estimateCommonDisp(d)
d<-estimateTagwiseDisp(d,prior.n=10)
exp.inf<-approx.expected.info(d,1/(1 + d$common.dispersion),d$pseudo.alt)

[Package edgeR version 2.4.3 Index]