as.data.frame {edgeR}	R Documentation

Turn a TopTags Object into a Dataframe

Description

Turn a TopTags object into a data.frame.

Usage

## S3 method for class 'TopTags'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

Arguments

x
an object of class TopTags

row.names
NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional
logical. If TRUE, setting row names and converting column names (to syntactic names) is optional.

...
additional arguments to be passed to or from methods.

Details

This method combines all the components of x which have a row for each tag (transcript) into a data.frame.

Value

A data.frame.

Author(s)

Gordon Smyth

See Also

as.data.frame in the base package.

[Package edgeR version 2.4.3 Index]