as.matrix {edgeR}	R Documentation

Turn a DGEList Object into a Matrix

Description

Turn a digital gene expression object into a numeric matrix by extracting the count values.

Usage

## S3 method for class 'DGEList'
as.matrix(x,...)

Arguments

x
an object of class DGEList.

...
additional arguments, not used for these methods.

Details

This method extracts the matrix of counts.

This involves loss of information, so the original data object is not recoverable.

Value

A numeric matrix.

Author(s)

Gordon Smyth

See Also

as.matrix in the base package or as.matrix.RGList in the limma package.

[Package edgeR version 2.4.3 Index]