commonCondLogLikDerDelta {edgeR}	R Documentation

Conditional Log-Likelihoods in Terms of Delta

Description

Common conditional log-likelihood parameterized in terms of delta (phi / (phi+1))

Usage

commonCondLogLikDerDelta(y, delta, der = 0, doSum = FALSE) 

Arguments

y
list with elements comprising the matrices of count data (or pseudocounts) for the different groups

delta
delta (phi / (phi+1)) parameter of negative binomial

der
derivative, either 0 (the function), 1 (first derivative) or 2 (second derivative)

doSum
logical, whether to sum over samples or not (default FALSE

Details

The common conditional log-likelihood is constructed by summing over all of the individual tag conditional log-likelihoods. The common conditional log-likelihood is taken as a function of the dispersion parameter (phi), and here parameterized in terms of delta (phi / (phi+1)). The value of delta that maximizes the common conditional log-likelihood is converted back to the phi scale, and this value is the estimate of the common dispersion parameter used by all tags.

Value

numeric scalar of function/derivative evaluated at given delta

Author(s)

Davis McCarthy

See Also

estimateCommonDisp is the user-level function for estimating the common dispersion parameter.

Examples

counts<-matrix(rnbinom(20,size=1,mu=10),nrow=5)
d<-DGEList(counts=counts,group=rep(1:2,each=2),lib.size=rep(c(1000:1001),2))
y<-splitIntoGroups(d)
ll1<-commonCondLogLikDerDelta(y,delta=0.5,der=0,doSum=FALSE)
ll2<-commonCondLogLikDerDelta(y,delta=0.5,der=1)

[Package edgeR version 2.4.3 Index]