condLogLikDerSize {edgeR}	R Documentation

Log-Likelihood of the Common Dispersion for a Single Equalized Group

Description

Derivatives of the conditional negative-binomial log-likelihood (for each tag/transcript) with respect to the common dispersion parameter, for a single group of replicate libraries of the same size. Parameterized in terms of size or precision (1/phi).

Usage

condLogLikDerSize(y, r, der=1)

Arguments

y
matrix of (pseudo) count data

r
size parameter of negative binomial distribution

der
order of derivative required, either 0 (the function), 1 (first derivative) or 2 (second derivative)

Details

The library sizes must be equalized before running this function. This function carries out the actual mathematical computations for the conditional log-likelihood and its derivatives, calculating the conditional log-likelihood for each tag/transcript.

Value

vector of function/derivative evaluations, one for each transcript

Author(s)

Mark Robinson, Davis McCarthy

Examples

y <- matrix(rnbinom(10,size=1,mu=10),nrow=5)
condLogLikDerSize(y,r=1,der=1)

[Package edgeR version 2.4.3 Index]