decideTestsDGE {edgeR}	R Documentation

Multiple Testing Across Genes and Contrasts

Description

Classify a series of related differential expression statistics as up, down or not significant. A number of different multiple testing schemes are offered which adjust for multiple testing down the genes as well as across contrasts for each gene.

Usage

decideTestsDGE(object, adjust.method="BH", p.value=0.05)

Arguments

object
deDGElist object, output from exactTest, or DGELRT object, output from DGELRT, from which p-values for differential expression and log-fold change values may be extracted.

adjust.method
character string specifying p-value adjustment method. Possible values are "none", "BH", "fdr" (equivalent to "BH"), "BY" and "holm". See p.adjust for details.

p.value
numeric value between 0 and 1 giving the desired size of the test

Details

These functions implement multiple testing procedures for determining whether each log-fold change in a matrix of log-fold changes should be considered significantly different from zero.

Value

An object of class TestResults (see TestResults). This is essentially a numeric matrix with elements -1, 0 or 1 depending on whether each DE p-value is classified as significant with negative log-fold change, not significant or significant with positive log-fold change, respectively.

Author(s)

Davis McCarthy, Gordon Smyth

See Also

Adapted from decideTests in the limma package.

[Package edgeR version 2.4.3 Index]