dispCoxReidSplineTrend {edgeR}	R Documentation

Estimate Dispersion Trend for Negative Binomial GLMs

Description

Estimate trended common dispersion parameters across multiple negative binomial generalized linear models using Cox-Reid adjusted profile likelihood.

Usage

dispCoxReidSplineTrend(y, design, offset=NULL, df = 5, subset=10000, method.optim="Nelder-Mead", trace=0)
dispCoxReidPowerTrend(y, design, offset=NULL, subset=10000, method.optim="Nelder-Mead", trace=0)

Arguments

y
numeric matrix of counts

design
numeric matrix giving the design matrix for the GLM that is to be fit.

offset
numeric scalar, vector or matrix giving the offset (in addition to the log of the effective library size) that is to be included in the NB GLM for the transcripts. If a scalar, then this value will be used as an offset for all transcripts and libraries. If a vector, it should be have length equal to the number of libraries, and the same vector of offsets will be used for each transcript. If a matrix, then each library for each transcript can have a unique offset, if desired. In adjustedProfileLik the offset must be a matrix with the same dimension as the table of counts.

df
integer giving the degrees of freedom of the spline function. The number of knots used for the spline function is df - 1.

subset
integer, number of rows to use in the calculation. Rows used are chosen evenly spaced by abundance.

method.optim
the method to be used in optim. See optim for more detail.

trace
logical, should iteration information be output?

Details

In the edgeR context, these are low-level functions called by estimateGLMTrendedDisp.

dispCoxReidSplineTrend maximizes the Cox-Reid adjusted profile likelihood (Cox and Reid, 1987) by fitting spline interpolation. dispCoxReidPowerTrend models the dispersion trend by a power function. The parameters of the power function are estimated by maximizing the Cox-Reid adjusted profile likelihood.

Value

Numeric vector giving the estimated trended common dispersions. It is of the same length as the number of tags in the count data.

Author(s)

Yunshun Chen, Davis McCarthy, Gordon Smyth

References

Cox, DR, and Reid, N (1987). Parameter orthogonality and approximate conditional inference. Journal of the Royal Statistical Society Series B 49, 1-39.

See Also

estimateGLMTrendedDisp, optim

Examples

design <- matrix(1,4,1)
y <- matrix((rnbinom(400,mu=100,size=2)),100,4)
dispCoxReidSplineTrend(y, design, df

=3)
dispCoxReidPowerTrend(y, design)

[Package edgeR version 2.4.3 Index]