estimateGLMCommonDisp {edgeR}	R Documentation

Estimate Common Dispersion for Negative Binomial GLMs

Description

Estimates a common negative binomial dispersion parameter for a DGE dataset with a general experimental design.

Usage

## S3 method for class 'DGEList'
estimateGLMCommonDisp(y, design=NULL, offset=NULL, method="CoxReid", ...)
## Default S3 method:
estimateGLMCommonDisp(y, design=NULL, offset=NULL, method="CoxReid", ...)

Arguments

y
an object that contains the raw counts for each library (the measure of expression level); it can either be a matrix of counts, or a DGEList object with (at least) elements counts (table of unadjusted counts) and samples (data frame containing information about experimental group, library size and normalization factor for the library size)

design
numeric matrix giving the design matrix for the GLM that is to be fit. Must be of full column rank. Defaults to a single column of ones, equivalent to treating the columns as replicate libraries.

method
method for estimating the dispersion. Possible values are "CoxReid", "Pearson" or "deviance".

offset
numeric scalar, vector or matrix giving the offsets for the log-linear models. If a scalar, then this value will be used as an offset for all transcripts and libraries. If a vector, it should be have length equal to the number of libraries, and the same vector of offsets will be used for each transcript. If a matrix, then it should have the same row and column dimensions as y.

In the DGEList method, the offset is calculated by default from the library sizes and normalization factors found in y$samples.

...
other arguments are passed to lower-level functions. See dispCoxReid, dispPearson and dispDeviance for details.

Details

This function calls dispCoxReid, dispPearson or dispDeviance depending on the method specified. See dispCoxReid for details of the three methods and a discussion of their relative performance.

Value

The default method returns a numeric vector of length 1 containing the estimated dispersion.

The DGEList method returns the same DGEList y as input but with common.dispersion as an added component.

Author(s)

Gordon Smyth

References

Robinson MD and Smyth GK (2008). Small-sample estimation of negative binomial dispersion, with applications to SAGE data. Biostatistics, 9, 321-332

See Also

dispCoxReid, dispPearson, dispDeviance

estimateGLMTrendedDisp for trended dispersion and estimateGLMTagwiseDisp for tagwise dispersions in the context of a generalized linear model.

estimateCommonDisp for common dispersion or estimateTagwiseDisp for tagwise dispersion in the context of a multiple group experiment (one-way layout).

Examples

#  True dispersion is 1/size=0.1
y <- matrix(rnbinom(1000,mu=10,size=10),ncol=4)
d <- DGEList(counts=y,group=c(1,1,2,2))
design <- model.matrix(~group, data=d$samples)
d1 <- estimateGLMCommonDisp(d, design)
d1$common.disp

#  Compare with classic CML estimator:
d2 <- estimateCommonDisp(d)
d2$common.disp

#  See example(glmFit) for a different example

[Package edgeR version 2.4.3 Index]