estimateGLMTrendedDisp {edgeR}	R Documentation

Estimate Trended Dispersion for Negative Binomial GLMs

Description

Estimates the dispersion parameter for each transcript (tag) with a trend that depends on the overall level of expression for the transcript for a DGE dataset for general experimental designs by using Cox-Reid approximate conditional inference for a negative binomial generalized linear model for each transcript (tag) with the unadjusted counts and design matrix provided.

Usage

## S3 method for class 'DGEList'
estimateGLMTrendedDisp(y, design, offset=NULL, method="bin.spline", ...)
## Default S3 method:
estimateGLMTrendedDisp(y, design, offset=NULL, method="bin.spline",  ...)

Arguments

y
an object that contains the raw counts for each library (the measure of expression level); it can either be a matrix of counts, or a DGEList object with (at least) elements counts (table of unadjusted counts) and samples (data frame containing information about experimental group, library size and normalization factor for the library size)

design
numeric matrix giving the design matrix for the GLM that is to be fit.

method
method (low-level function) used to estimated the trended dispersions. Possible values are "bin.spline" (default), "bin.loess" (which both result in a call to dispBinTrend), "power" (call to dispCoxReidPowerTrend), or "spline" (call to dispCoxReidSplineTrend).

offset
numeric scalar, vector or matrix giving the offset (in addition to the log of the effective library size) that is to be included in the NB GLM for the transcripts. If a scalar, then this value will be used as an offset for all transcripts and libraries. If a vector, it should be have length equal to the number of libraries, and the same vector of offsets will be used for each transcript. If a matrix, then each library for each transcript can have a unique offset, if desired. In adjustedProfileLik the offset must be a matrix with the same dimension as the table of counts. Default is NULL; if object is a DGEList and offset is NULL then offset will be calculated automatically from codey$samples.

...
other arguments are passed to lower-level functions. See dispBinTrend, dispCoxReidPowerTrend and dispCoxReidSplineTrend for details.

Details

This is a wrapper function for the lower-level functions that actually carry out the dispersion estimation calculations. Provide a convenient, object-oriented interface for users.

Value

When the input object is a DGEList, estimateGLMTrendedDisp produces a DGEList object, which contains the estimates of the trended dispersion parameter for the negative binomial model according to the method applied.

When the input object is a numeric matrix, the output of one of the lower-level functions dispBinTrend, dispCoxReidPowerTrend of dispCoxReidSplineTrend is returned.

Author(s)

Gordon Smyth, Davis McCarthy

References

Cox, DR, and Reid, N (1987). Parameter orthogonality and approximate conditional inference. Journal of the Royal Statistical Society Series B 49, 1-39.

See Also

dispBinTrend, dispCoxReidPowerTrend and dispCoxReidSplineTrend for details on how the calculations are done.

estimateGLMCommonDisp for common dispersion and estimateGLMTagwiseDisp for (trended) tagwise dispersion in the context of generalized linear models.

estimateCommonDisp for common dispersion or estimateTagwiseDisp for tagwise dispersion in the context of a multiple group experiment (one-way layout).

Examples

y <- matrix(rnbinom(1000,mu=10,size=10),ncol=4)
d <- DGEList(counts=y,group=c(1,1,2,2),lib.size=c(1000:1003))
design <- model.matrix(~group, data=d$samples) # Define the design matrix for the full model
disp <- estimateGLMTrendedDisp(d, design, min.n=10)

[Package edgeR version 2.4.3 Index]