estimatePs {edgeR}	R Documentation

Estimate Expression Levels

Description

Estimate expression levels (i.e. proportion of all sample mRNA corresponding to each tag; or, concentration of mRNA for each tag in sample mRNA) using maximum likelihood with dispersion parameter fixed based on the negative binomial model for each tag/gene and sample group. Expression proportions are used to determine overall abundance of each tag/gene and differential expression of tags/genes between groups.

Usage

estimatePs(object, r, tol = 1e-10, maxit = 30) 

Arguments

object
list containing (at least) the elements counts (table of counts), group (vector or factor indicating group) and lib.size (numeric vector of library sizes)

r
numeric vector providing the size parameter of negative binomial model (size = 1/phi where phi is the dispersion parameter in the NB model)

tol
numeric scalar, tolerance between iterations

maxit
positive integer scalar, maximum number of iterations

Details

The Newton-Raphson method is used to calculate iteratively the maximum likelihood estimate of the expression level (i.e. concentration of mRNA for a particular tag in the sample mRNA) for each tag/gene.

Value

A list with elements:

conc.common
numeric vector giving overall proportion/concentration for each tag

conc.group
numeric matrix with columns giving estimates of tag/gene concentrations (proportion of total RNA for that group that that particular tag/gene contributes) for different groups)

Author(s)

Mark Robinson, Davis McCarthy

Examples

set.seed(0)
y<-matrix(rnbinom(40,size=1,mu=10),ncol=4)
d<-DGEList(counts=y,group=rep(1:2,each=2),lib.size=rep(c(1000:1001),2))
conc<-estimatePs(d,r=1)

[Package edgeR version 2.4.3 Index]