estimateSmoothing {edgeR}	R Documentation

Estimate the Prior Weight for Tagwise Dispersions

Description

This function is no longer recommended or required. Use getPriorN instead.

Estimate the prior weight, prior.n, using an approximate empirical Bayes rule given the estimate of the common dispersion. The prior weight determines how much smoothing takes place to squeeze tag/genewise estimates of the dispersion closer to the estimate of the common dispersion.

Usage

estimateSmoothing(object,verbose=TRUE) 

Arguments

object
DGEList object, output of estimateCommonDisp

verbose
logical, whether to write comments, default true

Details

We are no longer recommending this function, as it produces variable results. prior.n is now set automatically using getPriorN.

Value

estimateSmoothing produces an object of class DGEList with the following components.

prior.n
scalar; estimate of the prior weight, i.e. the smoothing parameter that indicates the weight to put on the common likelihood compared to the individual tag's likelihood; prior.n of 10 means that the common likelihood is given 10 times the weight of the individual tag/gene's likelihood in the estimation of the tag/genewise dispersion

Author(s)

Mark Robinson, Davis McCarthy

See Also

getPriorN

Examples

y<-matrix(rnbinom(20,size=1,mu=10),nrow=5)
d<-DGEList(counts=y,group=rep(1:2,each=2),lib.size=rep(c(1000:1001),2))
d<-estimateCommonDisp(d)
prior.n<-estimateSmoothing(d)

[Package edgeR version 2.4.3 Index]