subsetting {edgeR}	R Documentation

Subset DGEList, DGEGLM, DGEExact and DGELRT Objects

Description

Extract a subset of a DGEList, DGEGLM, DGEExact or DGELRT object.

Usage

## S3 method for class 'DGEList'
object[i, j, ...]
## S3 method for class 'DGEGLM'
object[i, j, ...]
## S3 method for class 'DGEExact'
object[i, j, ...]
## S3 method for class 'DGELRT'
object[i, j, ...]

Arguments

object
object of class DGEList, DGEGLM, DGEExact or DGELRT, respectively

i,j
elements to extract. i subsets the tags or genes while j subsets the libraries. Note, columns of DGEGLM, DGEExact and DGELRT objects cannot be subsetted.

...
not used

Details

i,j may take any values acceptable for the matrix components of object of class DGEList. See the Extract help entry for more details on subsetting matrices. For DGEGLM, DGEExact and DGELRT objects, only rows (i.e. i) may be subsetted.

Value

An object of class DGEList, DGEGLM, DGEExact or DGELRT as appropriate, holding data from the specified subset of tags/genes and libraries.

Author(s)

Davis McCarthy, Gordon Smyth

See Also

Extract in the base package.

Examples

d <- matrix(rnbinom(16,size=1,mu=10),4,4)
rownames(d) <- c("a","b","c","d")
colnames(d) <- c("A1","A2","B1","B2")
d <- DGEList(counts=d,group=factor(c("A","A","B","B")))
d[1:2,]
d[1:2,2]
d[,2]
d <- estimateCommonDisp(d)
results <- exactTest(d)
results[1:2,]
# NB: cannot subset columns for DGEExact objects

[Package edgeR version 2.4.3 Index]