systematicSubset {edgeR}	R Documentation

Take a systematic subset of indices.

Description

Take a systematic subset of indices stratified by a ranking variable.

Usage

systematicSubset(n, order.by)

Arguments

n
integer giving the size of the subset.

order.by
numeric vector of the values by which the indices are ordered.

Value

systematicSubset returns a vector of size n.

Author(s)

Gordon Smyth

See Also

order

Examples

y <- rnorm(100, 1, 1)
systematicSubset(20, y)

[Package edgeR version 2.4.3 Index]